# backend/docs/swagger.yaml
definitions:
  models.NotaFiscal:
    properties:
      id:
        type: integer
      itens:
        items:
          $ref: '#/definitions/models.NotaFiscalItem'
        type: array
      numero_sequencial:
        type: integer
      status:
        type: string
    type: object
  models.NotaFiscalItem:
    properties:
      descricao_produto:
        type: string
      id:
        type: integer
      nota_id:
        type: integer
      produto_id:
        type: integer
      quantidade:
        type: integer
    type: object
  models.Produto:
    properties:
      codigo:
        type: string
      descricao:
        type: string
      id:
        type: integer
      saldo:
        type: integer
    type: object
info:
  contact: {}
paths:
  /faturamento:
    get:
      description: Retorna o histórico de notas fiscais com seus itens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.NotaFiscal'
            type: array
      summary: Listar todas as notas
      tags:
      - faturamento
    post:
      consumes:
      - application/json
      description: Cria uma nota fiscal, abate o estoque dos produtos e salva os itens
        em uma transação única.
      parameters:
      - description: Dados da Nota e Itens
        in: body
        name: nota
        required: true
        schema:
          $ref: '#/definitions/models.NotaFiscal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Abrir nova Nota Fiscal
      tags:
      - faturamento
  /faturamento/{id}/fechar:
    patch:
      description: Atualiza o status para Fechada e deduz o estoque dos produtos
      parameters:
      - description: ID da Nota Fiscal
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fechar e Imprimir Nota Fiscal
      tags:
      - faturamento
  /produtos:
    get:
      description: Retorna todos os produtos do banco de dados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Produto'
            type: array
      summary: Listar produtos
      tags:
      - produtos
    post:
      consumes:
      - application/json
      description: Salva um novo produto no banco de dados
      parameters:
      - description: Dados do Produto
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/models.Produto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Produto'
      summary: Cadastrar novo produto
      tags:
      - produtos
  /produtos/{id}/baixa:
    patch:
      consumes:
      - application/json
      description: Subtrai uma quantidade do saldo de um produto específico
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      - description: 'Quantidade a subtrair (ex: {'
        in: body
        name: quantidade
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Dar baixa no estoque
      tags:
      - produtos
swagger: "2.0"